test/Kriti-checked APIs
