+++ src/controller/favouriteGameController/index.js
@@ -1,0 +1,2 @@
+import { getFavouriteGame } from "./getFavouriteGame.js";
+export { getFavouriteGame };
