+++ src/controller/gameController/index.js
@@ -1,0 +1,7 @@
+import { getGameById } from "./getGamebyId.js";
+import { getGames } from "./getGames.js";
+import { getCurrentGame } from "./getCurrentGame.js";
+import { getGameHistory } from "./getGameHistory.js";
+import { placeBet, getValidBetOptions } from "./placeBet.js";
+
+export { getGameById, getGames,placeBet, getValidBetOptions ,getGameHistory, getCurrentGame};
