--- src/controller/gameDetailController/index.js
@@ -1,4 +1,0 @@
-import { getGameById } from "./getGamebyId.js";
-import { getGames } from "./getGames.js";
-
-export { getGameById, getGames };
