+++ src/controller/rulesController/index.js
@@ -1,0 +1,6 @@
+import { fetchRule } from "./fetchRule.js";
+import { deleteRule } from "./deleteRule.js";
+import { updateRule } from "./updateRule.js";
+import { createRule } from "./createRule.js";
+
+export { fetchRule, deleteRule, updateRule, createRule };
