+++ src/controller/walletController/index.js
@@ -1,0 +1,2 @@
+import { getWallet } from "./getWallet.js";
+export { getWallet };
