+++ src/routes/publicApiRoute.js
@@ -1,6 +1,7 @@
 import express from "express";
 import { loginUser, logoutUser } from "../controller/userController.js";
 import { registerUser } from "../controller/registrationController.js";
+import clientRoutesAPI from "./clientRoutesAPI.js";
 
 const router = express.Router();
 
@@ -16,5 +17,6 @@ router.post("/register", registerUser);
 router.post("/login", loginUser);
 router.post("/logout", logoutUser);
 
+router.use("/client", clientRoutesAPI);
 
 export default router;
